requirements.txt
SpeechRecognition
pyttsx3
pywhatkit
google-generativeai
PyMuPDF
requests
deep-translator
python-dotenv